refactor(l3b): move the virtual service object into the objects tree
