+++ bindings/go/dataplane_ut/dataplane_ut.go
@@ -13,6 +13,7 @@ package dataplaneut
 #cgo LDFLAGS: -L../../../build/modules/acl/dataplane
 #cgo LDFLAGS: -L../../../build/modules/fwstate/dataplane
 #cgo LDFLAGS: -L../../../build/objects/fwstate/api
+#cgo LDFLAGS: -L../../../build/objects/l3b/api
 #cgo LDFLAGS: -L../../../build/modules/forward/dataplane
 #cgo LDFLAGS: -L../../../build/modules/mirror/dataplane
 #cgo LDFLAGS: -L../../../build/modules/route/dataplane
@@ -41,7 +42,7 @@ package dataplaneut
 // references between them (fwstate->acl, worker->pipeline, etc.).
 // fwstate depends on acl — acl must come first inside the group.
 #cgo LDFLAGS: -Wl,--start-group
-#cgo LDFLAGS: -lblackhole_dp -ll3b_dp -ldecap_dp -ldscp_dp -lacl_dp -lfwstate_dp -lfwstate_objects -lforward_dp -lmirror_dp -lroute_dp -lroute_mpls_dp -lnat64_dp -lpdump_dp -lunrdup_dp
+#cgo LDFLAGS: -lblackhole_dp -ll3b_dp -ldecap_dp -ldscp_dp -lacl_dp -lfwstate_dp -lfwstate_objects -ll3b_objects -lforward_dp -lmirror_dp -lroute_dp -lroute_mpls_dp -lnat64_dp -lpdump_dp -lunrdup_dp
 #cgo LDFLAGS: -lplain_dp -lvlan_dp
 #cgo LDFLAGS: -ldataplane_ut -lpipeline -lmodule -lworker_dp -lconfig_dp -lpacket
 #cgo LDFLAGS: -llogging -lagent -lconfig_cp -lcounters -lerrors -lfilter_compiler -lfwstate -llib_utils
