+++ modules/l3b/api/meson.build
@@ -4,6 +4,7 @@ cp_dependencies = [
     lib_filter_compiler_dep,
     lib_config_cp_dep,
     lib_counters_dep,
+    lib_l3b_objects_dep,
 ]
 
 includes = include_directories('../dataplane')
