+++ modules/l3b/dataplane/config.h
@@ -3,109 +3,7 @@
 #include <lib/filter/filter.h>
 #include <lib/filter/rule.h>
 
-#include "common/network.h"
 #include "lib/controlplane/config/cp_module.h"
-#include "lib/controlplane/config/cp_object.h"
-
-enum real_state {
-	real_state_disabled = 0,
-	real_state_enabled = 1,
-};
-
-/*
- * A single backend (real server) reachable through an IP-in-IP tunnel.
- *
- * The address family of the tunnel is selected by type; source_net and
- * destination_addr carry the matching union member.
- */
-struct real_server {
-	// Outer tunnel address family (ip_family_ip4 or ip_family_ip6).
-	enum ip_family type;
-	// Source network used to derive the outer source address.
-	struct net source_net;
-	// Real server destination address used as the outer destination.
-	struct net_addr destination_addr;
-	// Whether the server is eligible to receive traffic.
-	enum real_state state;
-};
-
-/*
- * Source-side classification criteria of a virtual service.
- *
- * Each array is a relative pointer into module shared memory and is paired
- * with a matching count field. A packet matches when it falls into any of the
- * listed networks and port ranges.
- */
-struct source_filter {
-	uint32_t net6_count;
-	struct net6 *net6s;
-
-	uint32_t net4_count;
-	struct net4 *net4s;
-
-	uint32_t port_range_count;
-	struct filter_port_range *port_ranges;
-};
-
-/*
- * Round-robin index ring used by the scheduler to pick a real server.
- *
- * server_indexes is a relative pointer to an array of real_server array
- * indexes; size is the number of entries the ring holds.
- */
-struct real_ring {
-	uint32_t *server_indexes;
-	// Number of indexes currently populated; the dataplane selects among
-	// the first count entries.
-	uint32_t count;
-	// Maximum number of indexes the array can hold; set at configuration
-	// time.
-	uint32_t capacity;
-};
-
-/*
- * A virtual service exposed to clients.
- *
- * Incoming traffic that matches one of the per-family filters is dispatched
- * to a real server chosen through real_ring by a hash-derived scheduler.
- *
- * Contract: the controlplane must filter_init both filter_ip4 and filter_ip6
- * before publishing a virtual service; the dataplane queries them directly
- * (value_table_get assumes a non-NULL backing table).
- */
-struct virtual_service {
-	// Backends available for this service.
-	uint32_t real_server_count;
-	struct real_server *real_servers;
-
-	// Scheduler index ring over the real_servers array.
-	struct real_ring real_ring;
-
-	// Masks applied to the packet hash to derive a ring slot.
-	uint32_t scheduler_hash_mask;
-	uint32_t scheduler_index_mask;
-
-	// Per-family classification of incoming packets.
-	struct filter filter_ip6;
-	struct filter filter_ip4;
-};
-
-/*
- * A named virtual service published as a standalone shared-memory object.
- *
- * The cp_object header carries the (type, name) identity, the reference
- * accounting across configuration generations and the object's own memory
- * context backing every allocation below; module configurations reference
- * the service by name through cp_module_link_object and resolve it to the
- * embedded struct at execution-context build time.
- */
-struct l3b_virtual_service_object {
-	struct cp_object cp_object;
-	struct virtual_service virtual_service;
-};
-
-// Shared-memory object type under which virtual services are registered.
-#define L3B_VIRTUAL_SERVICE_OBJECT_TYPE "l3b_virtual_service"
 
 /*
  * Top-level l3b module configuration published into shared memory.
