+++ modules/l3b/dataplane/meson.build
@@ -3,11 +3,11 @@ dp_dependencies = [
   lib_filter_query_dep,
   lib_packet_dp_dep,
   lib_module_dp_dep,
+  lib_l3b_objects_dep,
 ]
 
 dp_sources = files(
   'dataplane.c',
-  'object.c',
 )
 
 lib_l3b_dp = static_library(
@@ -25,8 +25,5 @@ lib_l3b_dp_dep = declare_dependency(
     '-Wl,--defsym',
     '-Wl,new_module_l3b=new_module_l3b',
     '-Wl,--export-dynamic-symbol=new_module_l3b',
-    '-Wl,--defsym',
-    '-Wl,new_object_l3b_virtual_service=new_object_l3b_virtual_service',
-    '-Wl,--export-dynamic-symbol=new_object_l3b_virtual_service',
   ],
 )
