+++ objects/meson.build
@@ -1,1 +1,2 @@
 subdir('fwstate')
+subdir('l3b')
